requirements.txt
paho-mqtt==1.5.0
django
gunicorn
django-heroku
plotly